Reorganize code in modules
